+++ SistemaGian.Models/Producto.cs
@@ -29,7 +29,7 @@ public partial class Producto
     public string? Image { get; set; }
 
     public int? IdProveedor { get; set; }
-    public int? ProductoCantidad { get; set; }
+    public decimal? ProductoCantidad { get; set; }
 
     public int? Orden { get; set; }
